+++ config/mail.php
@@ -14,7 +14,7 @@ return [
     |
     */
 
-    'default' => env('MAIL_MAILER', 'log'),
+    'default' => env('MAIL_MAILER', 'failover'),
 
     /*
     |--------------------------------------------------------------------------
@@ -113,6 +113,11 @@ return [
     'from' => [
         'address' => env('MAIL_FROM_ADDRESS', 'hello@example.com'),
         'name' => env('MAIL_FROM_NAME', env('APP_NAME', 'Laravel')),
+    ],
+
+    'queue' => [
+        'connection' => env('MAIL_QUEUE_CONNECTION'),
+        'queue' => env('MAIL_QUEUE', 'mail'),
     ],
 
 ];